+++ crengine/include/lvfntman.h
@@ -20,6 +20,8 @@
 #include "lvstring.h"
 #include "lvstring32collection.h"
 #include "lvfont.h"
+#include "lvarray.h"
+#include "lvstream.h"
 
 /// font manager interface class
 class LVFontManager {
